consumer/requirements.txt
confluent_kafka
redis
psycopg2-binary
